refactor(wasm): split the demo page script into feature modules
The page script held the whole demo: the shared state, the settings dialog,
the playlist table, the detail panes and the scan flow. Relocate each into
its own module beside it, leaving demo.ts as the composition root that owns
the pick controls and wires the listeners. Behaviour and rendered output are
unchanged.

The state the modules share moves into one exported object, since an
imported binding is read-only at the import site. state.ts reads the stored
settings itself and imports no other page module at run time: the feature
modules import each other in cycles, so a module entered mid-evaluation
cannot safely call back into them while its own body runs.
